requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
fhir.resources>=7.0.0
httpx>=0.25.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
jinja2>=3.1.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
fhir.resources>=7.0.0
httpx>=0.25.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
jinja2>=3.1.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
PyYAML>=6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
bcrypt>=4.0.0
PyJWT>=2.8.0
email-validator>=2.0.0
